Create BTD.html
btd!!1!


fun fact: if you use tab to select an upgrade, you can spam the upgrade with enter (provided you have enough cash for it the first time). this will put you in debt but will also apply the effects of the upgrade every time you press enter. this is the most powerful with tack shooter's faster attack speed upgrade, as it lets it shoot at hypersonic speeds (the same as the super monkey). note that projectile life time exists, so more range upgrades won't do anything. this also works with the second game.